refactor Cart
